+++ hat/assets/js/apps/Iaso/domains/forms/components/FormLegendInput.tsx
@@ -10,7 +10,7 @@ import MESSAGES from '../messages';
 import {
     AddLegendDialog,
     EditLegendDialog,
-} from '../../../components/LegendBuilder/Dialog';
+} from '../../../components/LegendBuilder/LegendBuilkderDialog';
 import { Legend } from '../../../components/LegendBuilder/Legend';
 import { FormDataType } from '../types/forms';
 
@@ -43,78 +43,84 @@ export const FormLegendInput: FunctionComponent<FormFormProps> = ({
     const { formatMessage } = useSafeIntl();
     return (
         <>
-            {!isFormLoading && currentForm.legend_threshold?.value && (
-                <>
-                    {!isEmpty(currentForm.legend_threshold?.value) && (
-                        <Box position="relative" className={classes.fakeField}>
-                            <Grid container spacing={1}>
-                                <Grid item xs={6}>
-                                    <span className={classes.label}>
-                                        {formatMessage(MESSAGES.legend)}
-                                    </span>
-                                </Grid>
-                                <Grid
-                                    item
-                                    xs={6}
-                                    container
-                                    justifyContent="flex-end"
-                                    alignContent="flex-start"
-                                >
-                                    <EditLegendDialog
-                                        iconProps={{}}
-                                        titleMessage={MESSAGES.edit}
-                                        threshold={
-                                            currentForm.legend_threshold.value
-                                        }
-                                        onConfirm={newThreshold =>
-                                            setFieldValue(
-                                                'legend_threshold',
-                                                newThreshold,
-                                            )
-                                        }
-                                    />
-                                    <IconButton
-                                        size="small"
-                                        onClick={() =>
-                                            setFieldValue(
-                                                'legend_threshold',
-                                                null,
-                                            )
-                                        }
+            {!isFormLoading &&
+                currentForm.legend_threshold?.value !== undefined && (
+                    <>
+                        {!isEmpty(currentForm.legend_threshold?.value) && (
+                            <Box
+                                position="relative"
+                                className={classes.fakeField}
+                            >
+                                <Grid container spacing={1}>
+                                    <Grid item xs={6}>
+                                        <span className={classes.label}>
+                                            {formatMessage(MESSAGES.legend)}
+                                        </span>
+                                    </Grid>
+                                    <Grid
+                                        item
+                                        xs={6}
+                                        container
+                                        justifyContent="flex-end"
+                                        alignContent="flex-start"
                                     >
-                                        <DeleteIcon />
-                                    </IconButton>
+                                        <EditLegendDialog
+                                            iconProps={{}}
+                                            titleMessage={MESSAGES.edit}
+                                            threshold={
+                                                currentForm.legend_threshold
+                                                    .value
+                                            }
+                                            onConfirm={newThreshold =>
+                                                setFieldValue(
+                                                    'legend_threshold',
+                                                    newThreshold,
+                                                )
+                                            }
+                                        />
+                                        <IconButton
+                                            size="small"
+                                            onClick={() =>
+                                                setFieldValue(
+                                                    'legend_threshold',
+                                                    null,
+                                                )
+                                            }
+                                        >
+                                            <DeleteIcon />
+                                        </IconButton>
+                                    </Grid>
+                                    <Grid item xs={12}>
+                                        <Legend
+                                            threshold={
+                                                currentForm.legend_threshold
+                                                    .value
+                                            }
+                                        />
+                                    </Grid>
                                 </Grid>
-                                <Grid item xs={12}>
-                                    <Legend
-                                        threshold={
-                                            currentForm.legend_threshold.value
-                                        }
-                                    />
-                                </Grid>
-                            </Grid>
-                        </Box>
-                    )}
-                    {isEmpty(currentForm.legend_threshold.value) && (
-                        <Box
-                            display="flex"
-                            justifyContent="flex-end"
-                            alignContent="flex-start"
-                        >
-                            <AddLegendDialog
-                                iconProps={{}}
-                                titleMessage={MESSAGES.createLegend}
-                                onConfirm={newThreshold =>
-                                    setFieldValue(
-                                        'legend_threshold',
-                                        newThreshold,
-                                    )
-                                }
-                            />
-                        </Box>
-                    )}
-                </>
-            )}
+                            </Box>
+                        )}
+                        {isEmpty(currentForm.legend_threshold.value) && (
+                            <Box
+                                display="flex"
+                                justifyContent="flex-end"
+                                alignContent="flex-start"
+                            >
+                                <AddLegendDialog
+                                    iconProps={{}}
+                                    titleMessage={MESSAGES.createLegend}
+                                    onConfirm={newThreshold =>
+                                        setFieldValue(
+                                            'legend_threshold',
+                                            newThreshold,
+                                        )
+                                    }
+                                />
+                            </Box>
+                        )}
+                    </>
+                )}
         </>
     );
 };
